+++ src/GRA.Controllers/AvatarController.cs
@@ -68,7 +68,9 @@ namespace GRA.Controllers
                 Bundles = await _avatarService.GetUserUnlockBundlesAsync(),
                 PreconfiguredAvatars = await _avatarService.GetUserUnlockBundlesAsync(true),
                 DefaultLayer = userWardrobe.First(_ => _.DefaultLayer).Id,
-                ImagePath = _pathResolver.ResolveContentPath($"site{GetCurrentSiteId()}/avatars/")
+                ImagePath = _pathResolver.ResolveContentPath($"site{GetCurrentSiteId()}/avatars/"),
+                SharingEnabled =
+                    !await GetSiteSettingBoolAsync(SiteSettingKey.Avatars.DisableSharing)
             };
             if (user.PreconfiguredAvatarId.HasValue)
             {
@@ -113,6 +115,11 @@ namespace GRA.Controllers
 
         public async Task<IActionResult> Share()
         {
+            if (await GetSiteSettingBoolAsync(SiteSettingKey.Avatars.DisableSharing))
+            {
+                return RedirectToAction(nameof(Index));
+            }
+
             PageTitle = _sharedLocalizer[Annotations.Title.ShareYourAvatar];
             var userAvatar = await _avatarService.GetUserAvatarAsync();
             if (userAvatar?.Count > 0)
@@ -193,6 +200,11 @@ namespace GRA.Controllers
         [HttpPost]
         public async Task<IActionResult> Share(string selectionJson)
         {
+            if (await GetSiteSettingBoolAsync(SiteSettingKey.Avatars.DisableSharing))
+            {
+                return RedirectToAction(nameof(Index));
+            }
+
             if (!string.IsNullOrWhiteSpace(selectionJson))
             {
                 try
@@ -307,6 +319,11 @@ namespace GRA.Controllers
 
         public async Task<IActionResult> InstagramImage(string id)
         {
+            if (await GetSiteSettingBoolAsync(SiteSettingKey.Avatars.DisableSharing))
+            {
+                return RedirectToAction(nameof(Index));
+            }
+
             var siteId = GetCurrentSiteId();
 
             var igFilePath = _pathResolver.ResolveContentFilePath($"site{siteId}/igavatars/{id}.png");