+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
 ## Unreleased
 
+### Added
+
+- Notation on the participant dashboard that title and author are optional
+
 ## [4.2.1] - 2021-10-14
 
 ### Added